+++ packages/google_maps_flutter/google_maps_flutter_android/lib/src/google_maps_flutter_android.dart
@@ -229,6 +229,11 @@ class GoogleMapsFlutterAndroid extends GoogleMapsFlutterPlatform {
   }
 
   @override
+  Stream<MapPoiTapEvent> onPoiTap({required int mapId}) {
+    return _events(mapId).whereType<MapPoiTapEvent>();
+  }
+
+  @override
   Future<void> updateMapConfiguration(
     MapConfiguration configuration, {
     required int mapId,
@@ -1255,6 +1260,20 @@ class HostMapMessageHandler implements MapsCallbackApi {
   @override
   void onMarkerTap(String markerId) {
     streamController.add(MarkerTapEvent(mapId, MarkerId(markerId)));
+  }
+
+  @override
+  void onPoiTap(PlatformPointOfInterest poi) {
+    streamController.add(
+      MapPoiTapEvent(
+        mapId,
+        PointOfInterest(
+          position: LatLng(poi.position.latitude, poi.position.longitude),
+          name: poi.name,
+          placeId: poi.placeId,
+        ),
+      ),
+    );
   }
 
   @override