+++ packages/google_maps_flutter/google_maps_flutter_android/pigeons/messages.dart
@@ -235,6 +235,19 @@ class PlatformMarker {
   final String? clusterManagerId;
 }
 
+/// Pigeon equivalent of the Point of Interest class.
+class PlatformPointOfInterest {
+  PlatformPointOfInterest({
+    required this.position,
+    this.name,
+    required this.placeId,
+  });
+
+  final PlatformLatLng position;
+  final String? name;
+  final String placeId;
+}
+
 /// Pigeon equivalent of the Polygon class.
 class PlatformPolygon {
   PlatformPolygon({
@@ -805,6 +818,9 @@ abstract class MapsCallbackApi {
 
   /// Called when a marker cluster is tapped.
   void onClusterTap(PlatformCluster cluster);
+
+  /// Called when a POI is tapped.
+  void onPoiTap(PlatformPointOfInterest poi);
 
   /// Called when a polygon is tapped.
   void onPolygonTap(String polygonId);