+++ packages/google_maps_flutter/google_maps_flutter_ios/test/google_maps_flutter_ios_test.dart
@@ -1348,6 +1348,40 @@ void main() {
       reason: 'Should pass mapId on PlatformView creation message',
     );
   });
+
+  test('onPoiTap sends events to correct stream', () async {
+    const mapId = 1;
+    final fakePosition = PlatformLatLng(latitude: 12.34, longitude: 56.78);
+    const fakeName = 'Googleplex';
+    const fakePlaceId = 'iso_id_123';
+
+    final maps = GoogleMapsFlutterIOS();
+    // Initialize the handler which receives messages from the native side
+    final HostMapMessageHandler callbackHandler = maps.ensureHandlerInitialized(
+      mapId,
+    );
+
+    final stream = StreamQueue<MapPoiTapEvent>(maps.onPoiTap(mapId: mapId));
+
+    // Simulate a message from the native side via the Pigeon generated handler
+    callbackHandler.onPoiTap(
+      PlatformPointOfInterest(
+        position: fakePosition,
+        name: fakeName,
+        placeId: fakePlaceId,
+      ),
+    );
+
+    // Verify the event in the stream
+    final MapPoiTapEvent event = await stream.next;
+    expect(event.mapId, mapId);
+
+    final PointOfInterest poi = event.value;
+    expect(poi.position.latitude, fakePosition.latitude);
+    expect(poi.position.longitude, fakePosition.longitude);
+    expect(poi.name, fakeName);
+    expect(poi.placeId, fakePlaceId);
+  });
 }
 
 void _expectColorsEqual(PlatformColor actual, Color expected) {